[MIBM-322][SM]: Import and export emails - Country and currency still in English
