+++ conf/application.conf
@@ -62,7 +62,7 @@ play.http.secret.key = "HxPSfJTR0F07oO760QwLRiett9AWTRpAWWwqrsHYRwK3N0V0p8I48dDP
 
 # The application languages
 # ~~~~~
-play.i18n.langs = ["en"]
+play.i18n.langs = ["en", "cy"]
 
 # Router
 # ~~~~~
